Merging for render

+++ app/api/review_routes.py
@@ -17,6 +17,15 @@ def get_current_reviews():
     # print(" THIS IS ALL REVIEWS", new_reviews)
     return new_reviews
 
+@review_routes.route('/all')
+def get_all_reviews():
+    """
+    Gets all Reviews
+    """
+    all_reviews_obj = Review.query.all()
+    all_reviews = [review.to_dict() for review in all_reviews_obj]
+    return all_reviews
+
 @review_routes.route('/currentUser/<int:id>')
 @login_required
 def get_one_review(id):
@@ -35,13 +44,20 @@ def edit_review(id):
     Current user is able to get review by review id and edit
     """
     reviewObj = Review.query.get(id)
-    # review=reviewObj.to_dict()
-    form = ReviewForm()
-    reviewObj.review_text = form.data["review_text"]
-    reviewObj.star_rating = form.data["star_rating"]
-    db.session.commit()
 
-    return reviewObj.to_dict()
+    if not reviewObj:
+        return {"message":"This review does not exist"}
+
+    elif current_user.id == reviewObj.user_id:
+
+    # review=reviewObj.to_dict()
+        form = ReviewForm()
+        reviewObj.review_text = form.data["review_text"]
+        reviewObj.star_rating = form.data["star_rating"]
+        db.session.commit()
+
+        return reviewObj.to_dict()
+    return {"message": "This review does not belong to you"}
 
     # return review
 @review_routes.route('/delete/<int:id>', methods=['DELETE'])
@@ -51,6 +67,12 @@ def delete_review(id):
     Current user is able to delete review by review id
     """
     selected_review = Review.query.get(id)
-    db.session.delete(selected_review)
-    db.session.commit()
-    return 'DELETED'
+    if not selected_review:
+        return {"message":"This review does not exist"}
+
+    elif current_user.id == selected_review.user_id:
+
+        db.session.delete(selected_review)
+        db.session.commit()
+        return {"message": 'This has been deleted', "reviewId": id}
+    return {"message": "This review does not belong to you"}
